v0.6

+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: ochouikh <ochouikh@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/20 01:16:30 by ochouikh          #+#    #+#             */
-/*   Updated: 2023/05/13 14:48:08 by ochouikh         ###   ########.fr       */
+/*   Updated: 2023/05/22 21:24:38 by ochouikh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ typedef struct s_data
 	struct s_philo	*philos;
 }				t_data;
 
-int			ft_atoi(const char *str);
+long		ft_atoi(const char *str);
 void		*routine(void *arg);
 long long	current_time(void);
 void		my_usleep(long long time_to_sleep);
